et-backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1

# AI
google-generativeai==0.7.2

# Validation
pydantic==2.9.1
